Committing in ../doc  Modified Files:  	release.notes VeloDataProcessor bug fix digitise routine returned the number of ADC counts as an integer. Now changed to return a float, and taken the int when storing the charge of the VeloFullDigit, but kept the float for the noise.  Modified Files:  	VeloDataProcessor.cpp VeloDataProcessor.h
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@22848 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: e9ca73068c8c1a4af51f87372551a53f0aaa6dec

+++ Lbcom/Velo/VeloAlgorithms/src/VeloDataProcessor.cpp
@@ -84,7 +84,7 @@ StatusCode VeloDataProcessor::execute() {
     // take an MCFE make a VeloFullDigit
      VeloChannelID myKey((*MCFEIt)->key());
      VeloFullDigit* mydigit = new VeloFullDigit(myKey);
-     int ADC=digitise((*MCFEIt)->charge()); 
+     int ADC=int(digitise((*MCFEIt)->charge())); 
      mydigit->setRawADCValue(ADC);
 
      // following will come from data processing 
@@ -122,10 +122,10 @@ StatusCode VeloDataProcessor::finalize() {
     return StatusCode::SUCCESS;
 }
 
-int VeloDataProcessor::digitise(float electrons) {
+float VeloDataProcessor::digitise(float electrons) {
   // convert electrons to ADC counts
-  int digi = int(electrons*(VeloDigiParams::ADCFullScale/VeloDigiParams::electronsFullScale));
-  if (digi>(VeloDigiParams::ADCFullScale-1)) digi=int(VeloDigiParams::ADCFullScale-1);
+  float digi = electrons*(VeloDigiParams::ADCFullScale/VeloDigiParams::electronsFullScale);
+  if (digi>(VeloDigiParams::ADCFullScale-1)) digi=VeloDigiParams::ADCFullScale-1.;
 
   return digi;
 }
